Bug fix. Details below.
I had set a couple of key lines of code outside of an "if" statement which resulted in the background disappearing whenever you clicked on it, and no more words being rendered to the page - effectively stopping the game, with the timer still running.
It was a simple fix thankfully and now it should work fine. Deploying here to see how it looks live.